+++ agent/src/config/config.rs
@@ -658,6 +658,7 @@ pub struct YamlConfig {
     pub ntp_min_interval: Duration,
     pub l7_protocol_advanced_features: L7ProtocolAdvancedFeatures,
     pub oracle_parse_config: OracleParseConfig,
+    pub server_ports: Vec<u16>,
 }
 
 impl YamlConfig {
@@ -908,6 +909,10 @@ impl YamlConfig {
             c.l7_protocol_advanced_features.obfuscate_enabled_protocols = vec!["Redis".to_string()];
         }
 
+        if c.mirror_traffic_pcp > 9 {
+            c.mirror_traffic_pcp = 0;
+        }
+
         Ok(c)
     }
 
@@ -1092,6 +1097,7 @@ impl Default for YamlConfig {
                 resp_0x04_extra_byte: false,
             },
             ebpf_collector_queue_size: 65535,
+            server_ports: vec![],
         }
     }
 }